Changed test runner to pytest
This test runner plays nicely with PyCharm's debugger. Breakpoints now work and and detailed test output can be displayed within the IDE.
