+++ src/django_otp/plugins/otp_email/test_otp_email.py
@@ -7,7 +7,7 @@ from django.db import IntegrityError
 from django.test.utils import override_settings
 
 from django_otp.forms import OTPAuthenticationForm
-from django_otp.tests import TestCase, ThrottlingTestMixin
+from django_otp.test_otp import TestCase, ThrottlingTestMixin
 
 from .models import EmailDevice
