+++ tox.ini
@@ -9,12 +9,14 @@ envlist = static
 setenv = PYTHONPATH = {env:PYTHONPATH:}{:}{toxinidir}/test
          PYTHONWARNINGS = default
          DJANGO_SETTINGS_MODULE = test_project.settings
-deps = qrcode
+deps = pytest
+       pytest-django
+       qrcode
        freezegun
        django22: Django==2.2.*
        django30: Django==3.0.*
        django31: Django==3.1.*
-commands = {envbindir}/django-admin test django_otp
+commands = pytest --pyargs django_otp
 
 [testenv:static]
 basepython = python3
